refactor(web): split channel panel sub-components into their own files
Per the web STYLE_GUIDE ("one component per file"), assistant-channels-list.tsx
co-located five components — ChannelPanel, ConnectedChannelHeader,
ChannelTrustFloorSection, TelegramCredentialEntry, and TwilioCredentialEntry —
each with its own props and multi-branch rendering, not the trivial private
helper the co-location exception is reserved for.

Extract each into its own kebab-case file under domains/channels/components/,
and lift the shared CHANNEL_META registry into domains/channels/channel-meta.ts
so the list and the panel share one source instead of cross-importing.
assistant-channels-list.tsx keeps AssistantChannelsList and POLICY_CONFIRMATIONS.
No behavior change — the list's tests and stories are unchanged and still pass.
