+++ clients/web/src/domains/channels/components/assistant-channels-list.tsx
@@ -1,46 +1,24 @@
-import { CheckCircle } from "lucide-react";
 import { useEffect, useState } from "react";
 
-import { Button } from "@vellumai/design-library/components/button";
 import { ConfirmDialog } from "@vellumai/design-library/components/confirm-dialog";
-import { Dropdown } from "@vellumai/design-library/components/dropdown";
-import { Input } from "@vellumai/design-library/components/input";
-import { Notice } from "@vellumai/design-library/components/notice";
-import { Tag } from "@vellumai/design-library/components/tag";
-import { Typography } from "@vellumai/design-library/components/typography";
 
 import { DetailCard } from "@/components/detail-card";
-import { EmptyState } from "@/components/empty-state";
 import {
   MobileSidebarDrawer,
   MobileSidebarTrigger,
 } from "@/components/mobile-sidebar-drawer";
-import { assistantDisplayName as toAssistantDisplayName } from "@/utils/assistant-display-name";
+import type { MutationStatus } from "@/components/slack-setup-wizard";
 import { useChannelAdapterSelectionStore } from "@/domains/channels/adapter-selection-store";
+import { CHANNEL_META } from "@/domains/channels/channel-meta";
 import { ChannelAdapterList } from "@/domains/channels/components/channel-adapter-list";
-import { SlackChannelCard } from "@/domains/channels/components/slack-channel-card";
-import { SlackChannelSection } from "@/domains/channels/components/slack-channel-section";
-import { SlackConnectionCard } from "@/domains/channels/components/slack-connection-card";
-import {
-  SlackThreadBehavior,
-  type SlackThreadMode,
-} from "@/domains/channels/components/slack-thread-behavior";
-import {
-  SlackSetupWizard,
-  type MutationStatus,
-} from "@/components/slack-setup-wizard";
+import { ChannelPanel } from "@/domains/channels/components/channel-panel";
+import type { SlackThreadMode } from "@/domains/channels/components/slack-thread-behavior";
+import type { AdmissionPolicy } from "@/lib/channel-admission-policy/types";
 import type {
   AssistantChannelState,
   SetupChannelId,
 } from "@/types/channel-types";
-import {
-  ADMISSION_POLICY_DEFAULT,
-  ADMISSION_POLICY_VALUES,
-  getPolicyDescriptions,
-  POLICY_LABELS,
-  type AdmissionPolicy,
-} from "@/lib/channel-admission-policy/types";
-import { ChannelIcon, getChannelLabel } from "@/utils/channel-presentation";
+import { assistantDisplayName as toAssistantDisplayName } from "@/utils/assistant-display-name";
 
 type ChannelKey = SetupChannelId;
 
@@ -123,54 +101,15 @@ export interface AssistantChannelsListProps {
   initialChannel?: ChannelKey | null;
 }
 
-const CHANNEL_META: Record<
-  ChannelKey,
-  {
-    /** The disconnect-dialog subject ("Disconnect {label}?"). */
-    label: string;
-    disconnectMessage: string;
-    /**
-     * Whether a connected channel surfaces the "Who can message" trust-floor
-     * dropdown. Slack has none: its admission floors are managed per
-     * conversation type (DMs vs. channels), with no channel-wide knob.
-     */
-    hasTrustFloorControl: boolean;
-    /** One-line pitch for the disconnected empty state. Slack has none — its disconnected state is the setup wizard. */
-    disconnectedPitch?: (displayName: string) => string;
-  }
-> = {
-  slack: {
-    label: "Slack",
-    disconnectMessage:
-      "This clears the stored Slack bot and app tokens for this assistant. You can reconnect later.",
-    hasTrustFloorControl: false,
-  },
-  telegram: {
-    label: "Telegram",
-    disconnectMessage:
-      "This clears the stored Telegram bot token for this assistant. You can reconnect later.",
-    hasTrustFloorControl: true,
-    disconnectedPitch: (displayName) =>
-      `Connect a Telegram bot so ${displayName} can send and receive messages on Telegram.`,
-  },
-  phone: {
-    label: "Phone Calling",
-    disconnectMessage:
-      "This clears the stored Twilio credentials for this assistant. You can reconnect later.",
-    hasTrustFloorControl: true,
-    disconnectedPitch: (displayName) =>
-      `Connect your Twilio account so ${displayName} can make and answer phone calls.`,
-  },
-};
-
 /**
  * The Channels tab's master-detail surface: a left rail listing the
  * Slack/Telegram/Phone adapters (`ChannelAdapterList`) beside the selected
- * adapter's detail panel, plus the disconnect and trust-floor confirmation
- * dialogs. Rendered by the Channels tab (`ChannelsPage`). The active adapter
- * persists in `adapter-selection-store`; the queries and mutations behind the
- * props live in `useAssistantChannels`. The `channel-trust-floors` flag gates
- * the Channels tab itself (in `IntelligenceLayout`), not anything in here.
+ * adapter's detail panel (`ChannelPanel`), plus the disconnect and trust-floor
+ * confirmation dialogs. Rendered by the Channels tab (`ChannelsPage`). The
+ * active adapter persists in `adapter-selection-store`; the queries and
+ * mutations behind the props live in `useAssistantChannels`. The
+ * `channel-trust-floors` flag gates the Channels tab itself (in
+ * `IntelligenceLayout`), not anything in here.
  */
 export function AssistantChannelsList({
   assistantId,
@@ -360,441 +299,5 @@ export function AssistantChannelsList({
         onCancel={() => setPendingPolicy(null)}
       />
     </>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Channel Panel
-// ---------------------------------------------------------------------------
-
-interface ChannelPanelProps {
-  channel: AssistantChannelState;
-  assistantId: string;
-  assistantName: string;
-  /** Trimmed assistant name with a "your assistant" fallback, for copy. */
-  assistantDisplayName: string;
-  pending: boolean;
-  /**
-   * Open the manual credential form immediately instead of the empty state —
-   * set for `?setup=<channel>` deep links (e.g. the mobile chat-drawer
-   * handoff continues credential entry here).
-   */
-  initialManualEntry?: boolean;
-  onSetup?: () => void;
-  onDisconnect?: () => void;
-  onSaveTelegramToken?: (botToken: string) => Promise<void>;
-  onSaveSlackConfig?: (botToken: string, appToken: string) => void;
-  slackSaveStatus?: MutationStatus;
-  slackSaveError?: string | null;
-  slackThreadMode?: SlackThreadMode;
-  slackThreadModePending?: boolean;
-  onSlackThreadModeChange?: (mode: SlackThreadMode) => void;
-  onSaveTwilioCredentials?: (
-    accountSid: string,
-    authToken: string,
-  ) => Promise<void>;
-  policy?: AdmissionPolicy;
-  policySaving?: boolean;
-  policyLoading?: boolean;
-  policyError?: boolean;
-  onPolicyChange?: (policy: AdmissionPolicy) => void;
-}
-
-function ChannelPanel({
-  channel,
-  assistantId,
-  assistantName,
-  assistantDisplayName,
-  pending,
-  initialManualEntry = false,
-  onSetup,
-  onDisconnect,
-  onSaveTelegramToken,
-  onSaveSlackConfig,
-  slackSaveStatus,
-  slackSaveError,
-  slackThreadMode,
-  slackThreadModePending = false,
-  onSlackThreadModeChange,
-  onSaveTwilioCredentials,
-  policy,
-  policySaving = false,
-  policyLoading = false,
-  policyError = false,
-  onPolicyChange,
-}: ChannelPanelProps) {
-  const connected = channel.status === "ready";
-  // Manual credential entry is a connect-time affordance, so it only applies
-  // while disconnected — seeded from a `?setup=<channel>` deep link. Declared
-  // before the Slack branch to keep hook order stable across renders.
-  const [manualEntry, setManualEntry] = useState(initialManualEntry);
-
-  // Slack is its own adapter shape — a token-pair channel with dedicated
-  // connected/disconnected cards (connection card vs. setup wizard) that own
-  // their card chrome, so it returns bare (the parent skips the DetailCard).
-  if (channel.key === "slack") {
-    return (
-      <div className="flex flex-col gap-4">
-        {connected ? (
-          <SlackConnectionCard
-            slackHandle={channel.address}
-            disconnectPending={pending}
-            onDisconnect={onDisconnect}
-          >
-            <SlackThreadBehavior
-              threadMode={slackThreadMode}
-              threadModePending={slackThreadModePending}
-              onThreadModeChange={onSlackThreadModeChange}
-            />
-          </SlackConnectionCard>
-        ) : (
-          <SlackChannelCard>
-            <SlackSetupWizard
-              assistantName={assistantName}
-              onSave={onSaveSlackConfig}
-              saveStatus={slackSaveStatus}
-              saveError={slackSaveError}
-            />
-          </SlackChannelCard>
-        )}
-
-        {connected ? (
-          <SlackChannelSection
-            assistantId={assistantId}
-            assistantDisplayName={assistantDisplayName}
-            slackHandle={channel.address}
-          />
-        ) : null}
-      </div>
-    );
-  }
-
-  // Telegram and Phone are single-credential adapters that share one shape.
-  // Connected: the connection header plus the trust-floor control. Disconnected:
-  // a pitch with guided setup and a manual credential-entry escape hatch. The
-  // credential form is a connect-time affordance and never shows while
-  // connected — mirroring Slack's setup wizard, so "Connected" never sits next
-  // to an empty token field.
-  const meta = CHANNEL_META[channel.key];
-  return (
-    <div className="flex flex-col gap-4">
-      {connected ? (
-        <>
-          <ConnectedChannelHeader
-            address={channel.address}
-            pending={pending}
-            onDisconnect={onDisconnect}
-          />
-          {meta.hasTrustFloorControl && onPolicyChange ? (
-            <ChannelTrustFloorSection
-              assistantDisplayName={assistantDisplayName}
-              policy={policy}
-              saving={policySaving}
-              loading={policyLoading}
-              error={policyError}
-              onChange={onPolicyChange}
-            />
-          ) : null}
-        </>
-      ) : manualEntry ? (
-        channel.key === "telegram" ? (
-          <TelegramCredentialEntry onSave={onSaveTelegramToken} />
-        ) : (
-          <TwilioCredentialEntry onSave={onSaveTwilioCredentials} />
-        )
-      ) : (
-        // Pitch the channel and point at guided setup, with a
-        // manual-credentials escape hatch.
-        <EmptyState
-          icon={<ChannelIcon channelId={channel.key} className="h-6 w-6" />}
-          title={`${getChannelLabel(channel.key)} isn't connected`}
-          description={meta.disconnectedPitch?.(assistantDisplayName)}
-          action={
-            <div className="flex flex-col items-center gap-1">
-              <Button
-                type="button"
-                variant="outlined"
-                onClick={onSetup}
-                disabled={!onSetup || pending}
-              >
-                {pending ? "Opening…" : "Set up"}
-              </Button>
-              <Button
-                type="button"
-                variant="link"
-                onClick={() => setManualEntry(true)}
-              >
-                or connect manually
-              </Button>
-            </div>
-          }
-        />
-      )}
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Connected Channel Header (Telegram / Phone)
-// ---------------------------------------------------------------------------
-
-interface ConnectedChannelHeaderProps {
-  /** The connected channel's address/handle, when known. */
-  address?: string;
-  /** Disconnect in flight; disables the button and swaps its label. */
-  pending: boolean;
-  onDisconnect?: () => void;
-}
-
-/**
- * The connected-state header for the single-credential adapters (Telegram,
- * Phone): a Connected chip, the channel address, and a right-aligned
- * Disconnect affordance (the caller confirms first). Slack has its own
- * `SlackConnectionCard`; these channels render inside the panel's DetailCard.
- */
-function ConnectedChannelHeader({
-  address,
-  pending,
-  onDisconnect,
-}: ConnectedChannelHeaderProps) {
-  return (
-    <div className="flex items-center gap-3">
-      <Tag tone="positive" leftIcon={<CheckCircle />}>
-        Connected
-      </Tag>
-      {address ? (
-        <span
-          className="text-body-medium-lighter"
-          style={{ color: "var(--content-tertiary)" }}
-        >
-          {address}
-        </span>
-      ) : null}
-      <div className="ml-auto">
-        <Button
-          type="button"
-          variant="danger"
-          onClick={onDisconnect}
-          disabled={!onDisconnect || pending}
-        >
-          {pending ? "Disconnecting…" : "Disconnect"}
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Channel Trust Floor
-// ---------------------------------------------------------------------------
-
-interface ChannelTrustFloorSectionProps {
-  assistantDisplayName: string;
-  policy?: AdmissionPolicy;
-  saving?: boolean;
-  loading?: boolean;
-  error?: boolean;
-  onChange: (policy: AdmissionPolicy) => void;
-}
-
-function ChannelTrustFloorSection({
-  assistantDisplayName,
-  policy,
-  saving = false,
-  loading = false,
-  error = false,
-  onChange,
-}: ChannelTrustFloorSectionProps) {
-  const value = policy ?? ADMISSION_POLICY_DEFAULT;
-  const descriptions = getPolicyDescriptions(assistantDisplayName);
-  const options = ADMISSION_POLICY_VALUES.map((floor) => ({
-    value: floor,
-    label: POLICY_LABELS[floor],
-    tooltip: descriptions[floor],
-  }));
-
-  return (
-    <div className="flex flex-col gap-2">
-      <Typography
-        as="span"
-        variant="body-small-emphasised"
-        className="text-[color:var(--content-secondary)]"
-      >
-        Who can message {assistantDisplayName}
-      </Typography>
-      {loading ? (
-        // Hold off on rendering a concrete floor until the GET succeeds — the
-        // default would otherwise misreport a channel with a stored non-default
-        // (e.g. `no_one`) policy and let the user overwrite it.
-        <Typography
-          as="span"
-          variant="body-small-default"
-          className="text-[color:var(--content-tertiary)]"
-        >
-          Loading…
-        </Typography>
-      ) : error ? (
-        <Typography
-          as="span"
-          variant="body-small-default"
-          className="text-[color:var(--content-negative)]"
-        >
-          Couldn’t load this setting. Try reopening this page.
-        </Typography>
-      ) : (
-        <>
-          <div style={{ maxWidth: 280 }}>
-            <Dropdown<AdmissionPolicy>
-              value={value}
-              onChange={onChange}
-              options={options}
-              disabled={saving}
-              aria-label={`Who can message ${assistantDisplayName}`}
-            />
-          </div>
-          <Typography
-            as="span"
-            variant="body-small-default"
-            className="text-[color:var(--content-tertiary)]"
-          >
-            {descriptions[value]}
-          </Typography>
-          {value === "trusted_contacts" ? (
-            <Notice tone="info" className="max-w-lg">
-              People you haven’t verified yet — even teammates in the same
-              channel — can’t get through: {assistantDisplayName} lets them know
-              they need to be verified and notifies you. You can verify people
-              ahead of time in Contacts.
-            </Notice>
-          ) : null}
-        </>
-      )}
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Credential Entry Forms
-// ---------------------------------------------------------------------------
-
-interface TelegramCredentialEntryProps {
-  onSave?: (botToken: string) => Promise<void>;
-}
-
-function TelegramCredentialEntry({ onSave }: TelegramCredentialEntryProps) {
-  const [botToken, setBotToken] = useState("");
-  const [saving, setSaving] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const canSave = botToken.trim().length > 0 && !saving;
-
-  const handleSave = async () => {
-    if (!onSave || !canSave) {
-      return;
-    }
-    setSaving(true);
-    setError(null);
-    try {
-      await onSave(botToken.trim());
-      setBotToken("");
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "Failed to save");
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  return (
-    <div className="flex flex-col gap-3">
-      <Input
-        label="Bot Token"
-        type="password"
-        value={botToken}
-        onChange={(e) => setBotToken(e.target.value)}
-        placeholder="Paste your Telegram bot token"
-        disabled={saving}
-        fullWidth
-      />
-      {error ? (
-        <p
-          className="text-label-small"
-          style={{ color: "var(--content-negative)" }}
-        >
-          {error}
-        </p>
-      ) : null}
-      <div>
-        <Button type="button" onClick={handleSave} disabled={!canSave}>
-          {saving ? "Saving…" : "Save"}
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-interface TwilioCredentialEntryProps {
-  onSave?: (accountSid: string, authToken: string) => Promise<void>;
-}
-
-function TwilioCredentialEntry({ onSave }: TwilioCredentialEntryProps) {
-  const [accountSid, setAccountSid] = useState("");
-  const [authToken, setAuthToken] = useState("");
-  const [saving, setSaving] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const canSave =
-    accountSid.trim().length > 0 && authToken.trim().length > 0 && !saving;
-
-  const handleSave = async () => {
-    if (!onSave || !canSave) {
-      return;
-    }
-    setSaving(true);
-    setError(null);
-    try {
-      await onSave(accountSid.trim(), authToken.trim());
-      setAccountSid("");
-      setAuthToken("");
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "Failed to save");
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  return (
-    <div className="flex flex-col gap-3">
-      <Input
-        label="Account SID"
-        type="text"
-        value={accountSid}
-        onChange={(e) => setAccountSid(e.target.value)}
-        placeholder="ACxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
-        disabled={saving}
-        fullWidth
-      />
-      <Input
-        label="Auth Token"
-        type="password"
-        value={authToken}
-        onChange={(e) => setAuthToken(e.target.value)}
-        placeholder="Twilio auth token"
-        disabled={saving}
-        fullWidth
-      />
-      {error ? (
-        <p
-          className="text-label-small"
-          style={{ color: "var(--content-negative)" }}
-        >
-          {error}
-        </p>
-      ) : null}
-      <div>
-        <Button type="button" onClick={handleSave} disabled={!canSave}>
-          {saving ? "Saving…" : "Save"}
-        </Button>
-      </div>
-    </div>
   );
 }
